refactor: extract shared FileBuilder base to remove duplication
AddressFileBuilder and EstateFileBuilder were ~95% identical: the same
get/first/find/each/modify/delete methods copy-pasted across both, differing
only in the resource id enum, the parent-id parameter name, the relation type
and the id property.

Introduce an abstract FileBuilder holding the shared request logic, with four
small template methods (resourceId, parentIdParameter, relationType, parentId)
supplied by each concrete builder. Public constructors and properties
(addressId / estateId) are unchanged, so this is fully backward compatible.

Removes ~130 lines of duplicated code; full Pest suite (625 tests) and PHPStan
level 8 still pass.
